tool_resources/add-and-run-service/service-config-manager/template/requirements.txt
PyYAML==6.0.1
couchbase==4.3.2
kafka-python
psycopg2-binary==2.9.9
